pbrProject/requirements.txt
asgiref==3.8.1
beautifulsoup4==4.13.3
boto3==1.37.38
botocore==1.37.38
cachetools==5.5.2
certifi==2025.1.31
cffi==1.17.1
charset-normalizer==3.4.1
cryptography==44.0.2
dj-database-url==2.3.0
Django==4.2.19
django-allauth==65.7.0
django-bootstrap-v5==1.0.11
django-environ==0.12.0
django-storages==1.14.6
git-filter-repo==2.47.0
google-api-core==2.24.2
google-api-python-client==2.166.0
google-auth==2.38.0
google-auth-httplib2==0.2.0
googleapis-common-protos==1.69.2
gunicorn==23.0.0
httplib2==0.22.0
idna==3.10
jmespath==1.0.1
oauthlib==3.2.2
packaging==24.2
pillow==11.1.0
proto-plus==1.26.1
protobuf==6.30.2
psycopg2-binary==2.9.10
pyasn1==0.6.1
pyasn1_modules==0.4.2
pycparser==2.22
PyJWT==2.10.1
pyparsing==3.2.3
python-dateutil==2.9.0.post0
python-decouple==3.8
python-dotenv==1.1.0
requests==2.32.3
requests-oauthlib==2.0.0
rsa==4.9
s3transfer==0.11.5
six==1.17.0
soupsieve==2.6
sqlparse==0.5.3
typing_extensions==4.12.2
uritemplate==4.1.1
urllib3==2.3.0
whitenoise==6.9.0
